+++ cs/HomeExercises/Program.cs
@@ -1,0 +1,10 @@
+namespace HomeExercises
+{
+	// bug: Program does not contain a static 'Main' method suitable for an entry point
+	public class Program
+	{
+		public static void Main()
+		{
+		}
+	}
+}
